+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+
+  - Module application support added with docker-compose and files 
+    fields to be mapped with Docker secrets and configs
+
 ## [4.1.0] - 2019-07-11
 
 ### Changed